用户信息类完成

+++ Classes/util.cpp
@@ -1,6 +1,6 @@
 #include "Util.h"
 
-long long Util::getStringHash(const string& str)
+int Util::getStringHash(const string& str)
 {
 	const int length = str.length();
 	vector<int> hash(length + 10);
@@ -8,15 +8,15 @@ long long Util::getStringHash(const string& str)
 	hash[0] = length ? static_cast<long long>(str[0]) % MOD : 0;
 	for (auto i = 1; i < length; i++) hash[i] = (hash[i - 1] * BASE + static_cast<long long>(str[i])) % MOD;
 
-	return (hash[length - 1] - length * BASE + MOD) % MOD;
+	return static_cast<int>((hash[length - 1] - length * BASE + MOD) % MOD);
 }
 
-long long Util::getStringHash(vector<string> str)
+int Util::getStringHash(vector<string> str)
 {
 	sort(str.begin(), str.end());
 
 	long long hash = 0;
 	for (const auto& it : str) hash = (hash + getStringHash(it)) % MOD;
 
-	return hash;
+	return static_cast<int>(hash);
 }
